Implement multi-value charmaps (#1429)

+++ include/asm/section.hpp
@@ -84,9 +84,9 @@ void sect_EndUnion();
 void sect_CheckUnionClosed();
 
 void sect_AbsByte(uint8_t b);
-void sect_AbsByteString(std::vector<uint8_t> const &s);
-void sect_AbsWordString(std::vector<uint8_t> const &s);
-void sect_AbsLongString(std::vector<uint8_t> const &s);
+void sect_AbsByteString(std::vector<int32_t> const &s);
+void sect_AbsWordString(std::vector<int32_t> const &s);
+void sect_AbsLongString(std::vector<int32_t> const &s);
 void sect_Skip(uint32_t skip, bool ds);
 void sect_RelByte(Expression &expr, uint32_t pcShift);
 void sect_RelBytes(uint32_t n, std::vector<Expression> &exprs);
